refactor(perf): split /plan + /products into server shell + motion island
11th + 12th in the pattern series. Mirrors W17-W24.
